перенес миксин из файла с классом кейборд в модуль с классом айтем)))
